fixed error by separating auth into 2 files
